Expose Simfil Environment Constructor
Also rename files to match the naming schema of
the rest of the repo.
