+++ libs/model/src/simfilexpressioncache.h
@@ -15,7 +15,6 @@
 #include "simfil/model/model.h"
 #include "simfil/model/nodes.h"
 #include "simfil/simfil.h"
-#include "simfil-geometry.h"
 #include "simfil/expression.h"
 
 namespace mapget
@@ -26,18 +25,6 @@ namespace mapget
  * for a given node identifier.
  */
 using StringPoolResolveFun = std::function<std::shared_ptr<simfil::StringPool>(std::string_view const&)>;
-
-template <class... Args>
-std::unique_ptr<simfil::Environment> makeEnvironment(Args&& ...args)
-{
-    auto env = std::make_unique<simfil::Environment>(std::forward<Args>(args)...);
-    env->functions["geo"] = &mapget::GeoFn::Fn;
-    env->functions["point"] = &mapget::PointFn::Fn;
-    env->functions["bbox"] = &mapget::BBoxFn::Fn;
-    env->functions["linestring"] = &mapget::LineStringFn::Fn;
-
-    return env;
-}
 
 /**
  * Simfil compiled expression cache.
